coming

+++ resources/views/articles/show.blade.php
@@ -33,8 +33,10 @@
                 <div class="card-body">
                     <h5 class="card-title">{{$comment->name}}</h5>
                     <p class="card-text">{{$comment->desc}}</p>
-                    <a href="/comment/{{$comment->id}}/edit" class="btn btn-primary">Comment update</a>
-                    <a href="/comment/{{$comment->id}}/delete" class="btn btn-warning">Comment delete</a>
+                    @can('update_comment',$comment)
+                      <a href="/comment/{{$comment->id}}/edit" class="btn btn-primary">Comment update</a>
+                      <a href="/comment/{{$comment->id}}/delete" class="btn btn-warning">Comment delete</a>
+                    @endcan
                 </div>
                 </div>
             </div>
